work in progress

+++ app/Http/Controllers/PeriodController.php
@@ -3,12 +3,23 @@
 namespace App\Http\Controllers;
 
 use App\Models\Period;
+use App\Models\Teacher;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 class PeriodController extends Controller
 {
+
+
+    private static function validatePeriodData(Request $request, $required = true ) {
+        // same rule set but for update it's not required to fill all fields, and for store it does.
+        $rules = [
+            "name" => "string|max:255|unique:periods" . ($required ? "|required" : ""),
+            "teacher_id" => "int" . ($required ? "|required" : "")
+        ];
+        $request->validate($rules);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -36,9 +47,10 @@ class PeriodController extends Controller
      * @param Request $request
      * @return JsonResponse
      */
-    public function store(Request $request)
+    public function store(Request $request): JsonResponse
     {
-        $period = Period::create($request->all());
+        self::validatePeriodData($request);
+        $period = Period::create($request->only(['name','teacher_id']));
         return response()->json($period, 201);
     }
 
@@ -48,10 +60,13 @@ class PeriodController extends Controller
      * @param int $id
      * @return JsonResponse
      */
-    public function show(int $id)
+    public function show(int $id): JsonResponse
     {
-        $period = Period::findOrFail($id);
-        return response()->json($period);
+        $period = Period::find($id);
+        if ($period) {
+            return response()->json($period);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
     }
 
     /**
@@ -61,38 +76,66 @@ class PeriodController extends Controller
      * @param int $id
      * @return JsonResponse
      */
-    public function update(Request $request, int $id)
+    public function update(Request $request, int $id): JsonResponse
     {
-        $period = Period::findOrFail($id);
-        $period->update($request->all());
-        return response()->json($period, 200);
+        self::validatePeriodData($request,false);
+        $period = Period::find($id);
+        // because it's a foreign key, lets validate this teacher exists.
+        $teacher = Teacher::find($request->teacher_id);
+        if ($period && $teacher) {
+            $period->update($request->all());
+            return response()->json($period, 200);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param int $id
+     * @param string $id
      * @return JsonResponse
      */
-    public function destroy(int $id)
+    public function destroy(string $id): JsonResponse
     {
-        // Remove the specified period from the database
-        $period = Period::findOrFail($id);
-        $period->delete();
-        return response()->json(null, 204);
+        $period = Period::find($id);
+        if ($period) {
+            $period->delete();
+            return response()->json(null, 204);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
     }
 
-    public function addStudent(int $period_id, int $student_id)
+    /**
+     * Assign student to a period
+     *
+     * @param string $periodId
+     * @param string $studentId
+     * @return JsonResponse
+     */
+    public function addStudent(string $periodId, string $studentId): JsonResponse
     {
-        $period = Period::findOrFail($period_id);
-        $period->students()->attach($student_id);
-        return response()->json($period, 200);
-    }
+        $period = Period::find($periodId);
+        if ($period) {
+            $period->students()->attach($studentId);
+            return response()->json($period, 200);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
 
-    public function removeStudent(int $period_id, int $student_id)
+    }
+    /**
+     * Remove student from a period
+     *
+     * @param string $periodId
+     * @param string $studentId
+     * @return JsonResponse
+     */
+    public function removeStudent(string $periodId, string $studentId): JsonResponse
     {
-        $period = Period::findOrFail($period_id);
-        $period->students()->detach($student_id);
-        return response()->json($period, 200);
+        $period = Period::find($periodId);
+        if ($period) {
+            $period->students()->detach($studentId);
+            return response()->json($period, 200);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
     }
 }
