+++ app/Http/Controllers/TeacherController.php
@@ -1,11 +1,10 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\Http\Controllers\auth\TeacherAuthController;
 use App\Models\Teacher;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 
 class TeacherController extends Controller
 {
@@ -14,7 +13,7 @@ class TeacherController extends Controller
      *
      * @return JsonResponse
      */
-    public function index()
+    public function index(): JsonResponse
     {
         // Display a listing of teachers
         $teachers = Teacher::all();
@@ -37,49 +36,70 @@ class TeacherController extends Controller
      * @param Request $request
      * @return JsonResponse
      */
-    public function store(Request $request)
+    public function store(Request $request): JsonResponse
     {
-        $teacher = Teacher::create($request->all());
-        return response()->json($teacher, 201);
+        return app(TeacherAuthController::class)->register($request);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param int $id
+     * @param string $id
      * @return JsonResponse
      */
-    public function show(int $id)
+    public function show(string $id): JsonResponse
     {
-        $teacher = Teacher::findOrFail($id);
-        return response()->json($teacher);
+        $teacher = Teacher::find($id);
+        if ($teacher) {
+            return response()->json($teacher);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
+
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param Request $request
-     * @param  int  $id
+     * @param  string  $id
      * @return JsonResponse
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, string $id): JsonResponse
     {
-        $teacher = Teacher::findOrFail($id);
-        $teacher->update($request->all());
-        return response()->json($teacher, 200);
+        $user = $request->user();
+        if (strval($user->id) !== $id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+        $rules = TeacherAuthController::getValidation();
+        // same rules for reg / login but only refer to full_name
+        $rules = array_filter($rules, fn ($key) => in_array($key, ['full_name','email']), ARRAY_FILTER_USE_KEY);
+        $request->validate($rules);
+        $teacher = Teacher::find($id);
+        if (!$teacher) {
+            return response()->json(['error' => 'Resource not found'], 404);
+        }
+        $teacher->update($request->only(['full_name']));
+        return response()->json(['message' => 'Resource updated successfully', 'data' => $teacher]);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  string  $id
      * @return JsonResponse
      */
-    public function destroy($id)
+    public function destroy(string $id, Request $request): JsonResponse
     {
+        $user = $request->user();
+        if (strval($user->id) !== $id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         // Remove the specified teacher from the database
-        $teacher = Teacher::findOrFail($id);
-        $teacher->delete();
-        return response()->json(null, 204);
+        $teacher = Teacher::find($id);
+        if ($teacher) {
+            $teacher->delete();
+            return response()->json(null, 204);
+        }
+        return response()->json(['error' => 'Resource not found'], 404);
     }
 }
